Windows: Refactor pkg\plugins
